src/requirements.txt
numpy>=1.21.2,<1.24  
pandas>=1.3.3,<2.0.0
PyYAML>=6.0
scikit-learn>=1.0.1